[FIX/HM-170]: 방 상세 -> 공유 시 리다이렉트 적용
