+++ src/pages/RoomPage.tsx
@@ -4,6 +4,7 @@ import ConfirmList from '@/components/roomdetail/ConfirmList';
 import CreateNewMeeting from '@/components/roomdetail/CreateNewMeeting';
 import NonConfirmList from '@/components/roomdetail/NonConfirmList';
 import { PATH } from '@/constants/path';
+import { useRedirect } from '@/hooks/useRedirect';
 import useRoom from '@/hooks/useRoom';
 import useToolTip from '@/hooks/useToolTip';
 import {
@@ -17,15 +18,16 @@ import { toast } from 'sonner';
 import styled from 'styled-components';
 
 function RoomPage() {
+  useRedirect();
   const navigate = useNavigate();
-  const { id } = useParams();
+  const { roomId } = useParams();
   const { isToolTipOpen, closeToolTip } = useToolTip();
 
   const userInfo = sessionStorage.getItem('UserStore') || '';
   const parsedUserInfo = JSON.parse(userInfo);
   const userId = parsedUserInfo.state.user.id;
 
-  const { roomDetail, isError } = useRoom(Number(id));
+  const { roomDetail, isError } = useRoom(Number(roomId));
 
   if (isError) toast.error('잠시후 다시 시도해 주세요');
 
@@ -49,7 +51,7 @@ function RoomPage() {
   const handleCopyRoomUrl = async () => {
     try {
       await navigator.clipboard.writeText(
-        `${import.meta.env.VITE_APP_CLIENT_URL}/room/${id}`
+        `${import.meta.env.VITE_APP_CLIENT_URL}/room/${roomId}`
       );
       toast.info('링크가 복사되었습니다.');
     } catch (error) {
